Refactor XML import helpers
- move to common module
- update the default behavior and error messages a bit
- use the helpers both in XML-RPC and when manually importing
  test cases from XML
