+++ tcms/xmlrpc/api/testplan.py
@@ -9,6 +9,7 @@ from tcms.management.models import Product
 from tcms.testplans.models import TestPlan, TestPlanType, TCMSEnvPlanMap
 from tcms.xmlrpc.decorators import log_call
 from tcms.xmlrpc.utils import pre_process_ids, distinct_count
+from tcms.utils.xml import clean_xml_file
 
 __all__ = (
     'add_tag',
@@ -781,119 +782,3 @@ def import_case_via_XML(request, plan_id, values):
 
         tc.add_to_plan(plan=tp)
     return "Success update %d cases" % (i, )
-
-
-def clean_xml_file(xml_file):
-    from django.conf import settings
-    from tcms.core.contrib.xml2dict.xml2dict import XML2Dict
-
-    xml_file = xml_file.replace('\n', '')
-    xml_file = xml_file.replace('&testopia_', '&')
-    xml_file = xml_file.encode("utf8")
-
-    xml = XML2Dict()
-    xml_data = xml.fromstring(xml_file)
-    root_element = xml_data.get('testopia', None)
-    if root_element is None:
-        raise ValueError('Invalid XML document.')
-    if not root_element.get('version', None) != settings.TESTOPIA_XML_VERSION:
-        raise
-    case_elements = root_element.get('testcase', None)
-    if case_elements is not None:
-        if isinstance(case_elements, list):
-            return map(process_case, case_elements)
-        elif isinstance(case_elements, dict):
-            return map(process_case, (case_elements,))
-        else:
-            raise
-    else:
-        raise ValueError('No case found in XML document.')
-
-
-def process_case(case):
-    from django.contrib.auth.models import User
-    from tcms.management.models import Priority
-    from tcms.testcases.models import TestCaseStatus
-
-    # Check author
-    author = case.get('author', {}).get('value')
-    if author:
-        author = User.objects.get(email=author)
-        author_id = author.id
-    else:
-        raise ValueError('Invalid author: "{0}"'.format(author))
-
-    # Check default tester
-    default_tester_email = case.get('defaulttester', {}).get('value')
-    if default_tester_email:
-        default_tester = User.objects.get(email=default_tester_email)
-        default_tester_id = default_tester.id
-    else:
-        default_tester_id = None
-
-    # Check priority
-    priority = case.get('priority', {}).get('value')
-    if priority:
-        priority = Priority.objects.get(value=priority)
-        priority_id = priority.id
-    else:
-        raise ValueError('Invalid priority value: "{0}"'.format(priority))
-
-    # Check automated status
-    automated = case.get('automated', {}).get('value')
-    if automated:
-        is_automated = automated == 'Automatic' and True or False
-    else:
-        is_automated = False
-
-    # Check status
-    status = case.get('status', {}).get('value')
-    if status:
-        case_status = TestCaseStatus.objects.get(name=status)
-        case_status_id = case_status.id
-    else:
-        raise ValueError('Invalid status: "{0}"'.format(status))
-
-    # Check category
-    # *** Ugly code here ***
-    # There is a bug in the XML file, the category is related to product.
-    # But unfortunate it did not defined product in the XML file.
-    # So we have to define the category_name at the moment then get the product from the plan.
-    # If we did not found the category of the product we will create one.
-    category_name = case.get('categoryname', {}).get('value')
-    if not category_name:
-        raise ValueError('Invalid category name: "{0}"'.format(category_name))
-
-    # Check or create the tag
-    element = 'tag'
-    if case.get(element, {}):
-        tags = []
-        if isinstance(case[element], dict):
-            tag, create = TestTag.objects.get_or_create(name=case[element]['value'])
-            tags.append(tag)
-
-        if isinstance(case[element], list):
-            for tag_name in case[element]:
-                tag, create = TestTag.objects.get_or_create(name=tag_name['value'])
-                tags.append(tag)
-    else:
-        tags = None
-
-    new_case = {
-        'summary': case.get('summary', {}).get('value', ''),
-        'author_id': author_id,
-        'author': author,
-        'default_tester_id': default_tester_id,
-        'priority_id': priority_id,
-        'is_automated': is_automated,
-        'case_status_id': case_status_id,
-        'category_name': category_name,
-        'notes': case.get('notes', {}).get('value', ''),
-        'action': case.get('action', {}).get('value', ''),
-        'effect': case.get('expectedresults', {}).get('value', ''),
-        'setup': case.get('setup', {}).get('value', ''),
-        'breakdown': case.get('breakdown', {}).get('value', ''),
-        'tags': tags,
-    }
-
-    return new_case
